requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
# 可选：若使用 OpenAI 作为 LLM 后端
# openai>=1.0.0
